requirements.txt
Django==2.2.1
django-admin-rangefilter==0.4.0
django-taggit==0.22.2
django-widget-tweaks==1.4.3
Markdown==2.6.11
MonthDelta==0.9.1
mysqlclient==1.4.2.post1
numpy==1.16.4
pandas==0.24.2
psycopg2==2.7.4
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
